requirements.txt
django>=1.9.2
djangorestframework>=3.3
pillow
